Move 8-bit word into primitives
