+++ library/Rattletrap/AttributeValue/ClubColors.hs
@@ -1,6 +1,6 @@
 module Rattletrap.AttributeValue.ClubColors where
 
-import Rattletrap.Word8
+import Rattletrap.Primitive.Word8
 
 import qualified Data.Binary.Bits.Get as BinaryBit
 import qualified Data.Binary.Bits.Put as BinaryBit
