+++ library/Rattletrap/AttributeValue/Loadout.hs
@@ -1,7 +1,7 @@
 module Rattletrap.AttributeValue.Loadout where
 
 import Rattletrap.Word32
-import Rattletrap.Word8
+import Rattletrap.Primitive.Word8
 
 import qualified Data.Binary.Bits.Get as BinaryBit
 import qualified Data.Binary.Bits.Put as BinaryBit
