+++ library/Rattletrap/PropertyValue.hs
@@ -6,7 +6,7 @@ import Rattletrap.Primitive.Int32
 import Rattletrap.Primitive.List
 import Rattletrap.Primitive.Text
 import Rattletrap.Word64
-import Rattletrap.Word8
+import Rattletrap.Primitive.Word8
 
 import qualified Data.Binary as Binary
