+++ library/Rattletrap/RemoteId.hs
@@ -1,7 +1,7 @@
 module Rattletrap.RemoteId where
 
 import Rattletrap.Word64
-import Rattletrap.Word8
+import Rattletrap.Primitive.Word8
 
 import qualified Data.Binary.Bits.Get as BinaryBit
 import qualified Data.Binary.Bits.Put as BinaryBit
